# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get the currently authenticated user profile",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update the currently authenticated user profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/me/exploration-context": {
      "patch": {
        "operationId": "UsersController_updateExplorationContext",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExplorationContextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update the active exploration context for the user",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/business": {
      "post": {
        "operationId": "BusinessProfileController_createProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "business management"
        ]
      }
    },
    "/api/business/{id}": {
      "patch": {
        "operationId": "BusinessProfileController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "business management"
        ]
      },
      "delete": {
        "operationId": "BusinessProfileController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "business management"
        ]
      }
    },
...
    "/api/business/{id}/dashboard-stats": {
      "get": {
        "operationId": "BusinessProfileController_getStats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "business management"
        ]
      }
    },
    "/api/business/{id}/hours": {
      "put": {
        "operationId": "BusinessProfileController_updateHours",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetOperatingHoursDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "business management"
        ]
      }
    },
    "/api/business/{id}/tags": {
      "put": {
        "operationId": "BusinessProfileController_updateTags",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTagsDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "business management"
        ]
      }
    },
...
    "/api/v1/locations/search": {
      "get": {
        "operationId": "LocationsController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "Search query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Search for locations",
        "tags": [
          "Locations"
        ]
      }
    },
    "/api/v1/locations/reverse": {
      "get": {
        "operationId": "LocationsController_reverse",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "description": "Latitude",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lon",
            "required": true,
            "in": "query",
            "description": "Longitude",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Reverse geocode coordinates",
        "tags": [
          "Locations"
        ]
      }
    },
    "/api/v1/locations/ensure": {
      "post": {
        "operationId": "LocationsController_ensure",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnsureLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Ensure a location is persisted from geocoder results",
        "tags": [
          "Locations"
        ]
      }
    },
    "/api/v1/locations/{id}": {
      "get": {
        "operationId": "LocationsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a location by ID",
        "tags": [
          "Locations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Orita API",
    "description": "Local discovery platform connecting people with businesses, services, and opportunities around them",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateUserProfileDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateExplorationContextDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateBusinessProfileDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateBusinessProfileDto": {
        "type": "object",
        "properties": {}
      },
      "SetOperatingHoursItemDto": {
        "type": "object",
        "properties": {
          "day": {
            "type": "string",
            "enum": [
              "MON",
              "TUE",
              "WED",
              "THU",
              "FRI",
              "SAT",
              "SUN"
            ]
          },
          "openTime": {
            "type": "string",
            "example": "09:00"
          },
          "closeTime": {
            "type": "string",
            "example": "17:00"
          },
          "isClosed": {
            "type": "boolean"
          }
        },
        "required": [
          "day",
          "openTime",
          "closeTime",
          "isClosed"
        ]
      },
      "SetOperatingHoursDto": {
        "type": "object",
        "properties": {
          "hours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetOperatingHoursItemDto"
            }
          }
        },
        "required": [
          "hours"
        ]
      },
      "SetTagsDto": {
        "type": "object",
        "properties": {
          "tagIds": {
            "description": "Array of Tag IDs to assign to the business",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "tagIds"
        ]
      },
...
      "EnsureLocationDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}